+++ ShortGame1/DebugMapGraphicsService.h
@@ -37,11 +37,11 @@ namespace GEM
 
 		virtual void shutdown() override;
 
-		virtual ActionResult preFrame(double timeDelta) override;
+		virtual ActionResult preFrame(float timeDelta) override;
 
-		virtual ActionResult frame(double timeDelta) override;
+		virtual ActionResult frame(float timeDelta) override;
 
-		virtual ActionResult postFrame(double timeDelta) override;
+		virtual ActionResult postFrame(float timeDelta) override;
 
 
 		//Inherited from SDL_MouseListener
